Show confirmation for delete setting
Issue: BMS-3784 / BMS-3786

+++ src/main/webapp/WEB-INF/static/js/import-crosses.js
@@ -310,6 +310,9 @@ var ImportCrosses = {
 			if (!(data && data.programPresetId)) {
 				return;
 			}
+			if (!confirm(crossingSettingsDeleteConfirm)) {
+				return;
+			}
 			ImportCrosses.deleteImportSettings(data.programPresetId)
 				.done(function () {
 					showSuccessfulMessage('', crossingSettingsDeleted);
